refact custom nav
